+++ api/src/no/mnemonic/services/grafeo/api/service/v1/GrafeoService.java
@@ -355,63 +355,6 @@ public interface GrafeoService extends Service {
   }
 
   /**
-   * Traverse the graph of Objects and Facts starting at an Object identified by its ID.
-   *
-   * @param rh      Contains meta data about the request.
-   * @param request Request containing graph traversal query.
-   * @return Result of the graph traversal.
-   * @throws AccessDeniedException         If the user is not allowed to perform this operation.
-   * @throws AuthenticationFailedException If the user could not be authenticated.
-   * @throws InvalidArgumentException      If the request contains invalid parameters.
-   * @throws OperationTimeoutException     If the graph traversal timed out.
-   *
-   * @deprecated see {@link #traverse(RequestHeader, TraverseGraphByObjectsRequest)} instead
-   */
-  @Deprecated
-  default ResultSet<?> traverseGraph(RequestHeader rh, TraverseByObjectIdRequest request)
-          throws AccessDeniedException, AuthenticationFailedException, InvalidArgumentException, OperationTimeoutException {
-    throw new UnsupportedOperationException();
-  }
-
-  /**
-   * Traverse the graph of Objects and Facts starting at an Object identified by its type and value.
-   *
-   * @param rh      Contains meta data about the request.
-   * @param request Request containing graph traversal query.
-   * @return Result of the graph traversal.
-   * @throws AccessDeniedException         If the user is not allowed to perform this operation.
-   * @throws AuthenticationFailedException If the user could not be authenticated.
-   * @throws InvalidArgumentException      If the request contains invalid parameters.
-   * @throws OperationTimeoutException     If the graph traversal timed out.
-   *
-   * @deprecated see {@link #traverse(RequestHeader, TraverseGraphByObjectsRequest)} instead
-   */
-  @Deprecated
-  default ResultSet<?> traverseGraph(RequestHeader rh, TraverseByObjectTypeValueRequest request)
-          throws AccessDeniedException, AuthenticationFailedException, InvalidArgumentException, OperationTimeoutException {
-    throw new UnsupportedOperationException();
-  }
-
-  /**
-   * Traverse the graph of Objects and Facts starting at the Objects returned from an Object search.
-   *
-   * @param rh      Contains meta data about the request.
-   * @param request Request containing graph traversal query and Object search parameters.
-   * @return Result of the graph traversal.
-   * @throws AccessDeniedException         If the user is not allowed to perform this operation.
-   * @throws AuthenticationFailedException If the user could not be authenticated.
-   * @throws InvalidArgumentException      If the request contains invalid parameters.
-   * @throws OperationTimeoutException     If the graph traversal timed out.
-   *
-   * @deprecated see {@link #traverse(RequestHeader, TraverseGraphByObjectSearchRequest)} instead
-   */
-  @Deprecated
-  default ResultSet<?> traverseGraph(RequestHeader rh, TraverseByObjectSearchRequest request)
-          throws AccessDeniedException, AuthenticationFailedException, InvalidArgumentException, OperationTimeoutException {
-    throw new UnsupportedOperationException();
-  }
-
-  /**
    * Traverse the graph of Objects and Facts starting at a set of Objects
    *
    * @param rh      Contains meta data about the request.